+++ dtslint/ts4.1/ReadonlyRecord.ts
@@ -70,12 +70,23 @@ _.collect((_k: 'a' | 'b', n: number) => n)(r1) // $ExpectType readonly number[]
 // insertAt
 //
 
-_.insertAt('b', 0)({ a: 1 }) // $ExpectType Readonly<Record<"a" | "b", number>>
-_.insertAt('b', 0)(l1) // $ExpectType Readonly<Record<"a" | "b", number>>
-_.insertAt('b', 0)(d1) // $ExpectType Readonly<Record<string, number>>
-_.insertAt('b', 0)(r1) // $ExpectType Readonly<Record<"a" | "b", number>>
-_.insertAt(stringKey, 0)(r1) // $ExpectType Readonly<Record<string, number>>
-_.insertAt('c', 0)(r1) // $ExpectType Readonly<Record<"a" | "b" | "c", number>>
+_.insertAt('b', 0)({ a: 1 }) // $ExpectType Option<Readonly<Record<"a" | "b", number>>>
+_.insertAt('b', 0)(l1) // $ExpectType Option<Readonly<Record<"a" | "b", number>>>
+_.insertAt('b', 0)(d1) // $ExpectType Option<Readonly<Record<string, number>>>
+_.insertAt('b', 0)(r1) // $ExpectType Option<Readonly<Record<"a" | "b", number>>>
+_.insertAt(stringKey, 0)(r1) // $ExpectType Option<Readonly<Record<string, number>>>
+_.insertAt('c', 0)(r1) // $ExpectType Option<Readonly<Record<"a" | "b" | "c", number>>>
+
+//
+// upsertAt
+//
+
+_.upsertAt('b', 0)({ a: 1 }) // $ExpectType Readonly<Record<"a" | "b", number>>
+_.upsertAt('b', 0)(l1) // $ExpectType Readonly<Record<"a" | "b", number>>
+_.upsertAt('b', 0)(d1) // $ExpectType Readonly<Record<string, number>>
+_.upsertAt('b', 0)(r1) // $ExpectType Readonly<Record<"a" | "b", number>>
+_.upsertAt(stringKey, 0)(r1) // $ExpectType Readonly<Record<string, number>>
+_.upsertAt('c', 0)(r1) // $ExpectType Readonly<Record<"a" | "b" | "c", number>>
 
 //
 // deleteAt